Modernize the project build and add missing files
This change migrates the project from a Makefile to Meson, so that
dependencies can be intelligently checked and it is easy to package
the application for Linux distributions.

Additionally, the README is revised with correct instructions on
building and installing the code. The formatting of the file is also
slightly cleaned up. The duplicate copy of the license is removed.

Also, the missing Code of Conduct file is added to the project.
A gitignore file is also added to reduce the risk of accidentally
committing unwanted files.

Finally, basic GitHub Actions CI to test the build is added to
ensure the project compiles as new contributions come in.

Signed-off-by: Neal Gompa <neal@gompa.dev>